Modify tests into one file addremove.test.js
